+++ samples/ProcessPercent.ps1
@@ -9,8 +9,8 @@ Select-Object -First 25 |
 Format-Table -Property Name, ID, Handles,
 @{Name = 'WS(MB)'; Expression = { Format-Value $_.WS -Unit MB } },
 @{Name = 'PctWS'; Expression = {
-        $pct = Format-Percent -Value $_.ws -Total $total -Decimal 2
-        $bar = New-ANSIBar -Range 14 -Spacing $pct -Character BlackSquare
-        '{0:00.00} {1}' -f $pct, $bar
-    }
+    $pct = Format-Percent -Value $_.ws -Total $total -Decimal 2
+    $bar = New-ANSIBar -Range 14 -Spacing $pct -Character BlackSquare
+    '{0:00.00} {1}' -f $pct, $bar
+}
 }